Crate vst2

rust-vst2 is a rust implementation of the VST2.4 API

Plugins

All Plugins must implement the Plugin trait and std::default::Default. The plugin_main! macro must also be called in order to export the necessary functions for the plugin to function.

Plugin Trait

All methods in this trait have a default implementation except for the get_info method which must be implemented by the plugin. Any of the default implementations may be overriden for custom functionality; the defaults do nothing on their own.

plugin_main! macro

plugin_main! will export the necessary functions to create a proper VST plugin. This must be called with your VST plugin struct name in order for the vst to work.

Example plugin

A barebones VST plugin:

#[macro_use]
extern crate vst2;

use vst2::plugin::{Info, Plugin};

#[derive(Default)]
struct BasicPlugin;

impl Plugin for BasicPlugin {
    fn get_info(&self) -> Info {
        Info {
            name: "Basic Plugin".to_string(),
            unique_id: 1357, // Used by hosts to differentiate between plugins.

            ..Default::default()
        }
    }
}

plugin_main!(BasicPlugin); // Important!

Hosts

Hosts are currently not supported. TODO

Modules

api
Structures and types for interfacing with the VST 2.4 API.

buffer
Buffers to safely work with audio samples.

channels
Meta data for dealing with input / output channels. Not all hosts use this so it is not necessary for plugin functionality.

editor
All VST plugin editor related functionality.

host
Wrapper around host callback for plugins.

plugin
Plugin specific structures.

Macros

plugin_main!
Exports the necessary symbols for the plugin to be used by a VST host.