Enum vst2::api::SpeakerArrangementType

pub enum SpeakerArrangementType {
    Custom,
    Empty,
    Mono,
    Stereo,
    StereoSurround,
    StereoCenter,
    StereoSide,
    StereoCLfe,
    Cinema30,
    Music30,
    Cinema31,
    Music31,
    Cinema40,
    Music40,
    Cinema41,
    Music41,
    Surround50,
    Surround51,
    Cinema60,
    Music60,
    Cinema61,
    Music61,
    Cinema70,
    Music70,
    Cinema71,
    Music71,
    Cinema80,
    Music80,
    Cinema81,
    Music81,
    Surround102,
}

Tells the host how the channels are intended to be used in the plugin. Only useful for some hosts.

Variants

Custom

User defined arrangement.

Empty

Empty arrangement.

Mono

Mono.

Stereo

L R

StereoSurround

Ls Rs

StereoCenter

Lc Rc

StereoSide

Sl Sr

StereoCLfe

C Lfe

Cinema30

L R C

Music30

L R S

Cinema31

L R C Lfe

Music31

L R Lfe S

Cinema40

L R C S (LCRS)

Music40

L R Ls Rs (Quadro)

Cinema41

L R C Lfe S (LCRS + Lfe)

Music41

L R Lfe Ls Rs (Quadro + Lfe)

Surround50

L R C Ls Rs

Surround51

L R C Lfe Ls Rs

Cinema60

L R C Ls Rs Cs

Music60

L R Ls Rs Sl Sr

Cinema61

L R C Lfe Ls Rs Cs

Music61

L R Lfe Ls Rs Sl Sr

Cinema70

L R C Ls Rs Lc Rc

Music70

L R C Ls Rs Sl Sr

Cinema71

L R C Lfe Ls Rs Lc Rc

Music71

L R C Lfe Ls Rs Sl Sr

Cinema80

L R C Ls Rs Lc Rc Cs

Music80

L R C Ls Rs Cs Sl Sr

Cinema81

L R C Lfe Ls Rs Lc Rc Cs

Music81

L R C Lfe Ls Rs Cs Sl Sr

Surround102

L R C Lfe Ls Rs Tfl Tfc Tfr Trl Trr Lfe2