Module vst2::api::flags

Bitflags.

Structs

Channel
Flags for VST channels.

MidiEvent
MIDI event flags.

ModifierKey
Cross platform modifier key flags.

Plugin
Flags for VST plugins.

Constants

ACTIVE
Indicates channel is active. Ignored by host.

ALT
Alt key.

CAN_DOUBLE_REPLACING
Supports 64 bit audio processing.

CAN_REPLACING
Plugin can process 32 bit audio. (Mandatory in VST 2.4).

COMMAND
Control on mac.

CONTROL
Command on mac, ctrl on other.

HAS_EDITOR
Plugin has an editor.

IS_SYNTH
Plugin is a synth.

NO_SOUND_IN_STOP
Plugin does not produce sound when all input is silence.

PROGRAM_CHUNKS
Plugin preset data is handled in formatless chunks.

REALTIME_EVENT
This event is played live (not in playback from a sequencer track). This allows the plugin to handle these flagged events with higher priority, especially when the plugin has a big latency as per plugin::Info::initial_delay.

SHIFT
Shift key.

SPEAKER
Use channel's specified speaker_arrangement instead of stereo flag.

STEREO
Indicates channel is first of stereo pair.