Struct vst2::api::AEffect

pub struct AEffect {
    pub magic: i32,
    pub dispatcher: DispatcherProc,
    pub _process: ProcessProc,
    pub setParameter: SetParameterProc,
    pub getParameter: GetParameterProc,
    pub numPrograms: i32,
    pub numParams: i32,
    pub numInputs: i32,
    pub numOutputs: i32,
    pub flags: i32,
    pub reserved1: isize,
    pub reserved2: isize,
    pub initialDelay: i32,
    pub _realQualities: i32,
    pub _offQualities: i32,
    pub _ioRatio: f32,
    pub object: *mut c_void,
    pub user: *mut c_void,
    pub uniqueId: i32,
    pub version: i32,
    pub processReplacing: ProcessProc,
    pub processReplacingF64: ProcessProcF64,
    pub future: [u8; 56],
}

Used with the VST API to pass around plugin information.

Fields

magic

Magic number. Must be ['V', 'S', 'T', 'P'].

dispatcher

Host to plug-in dispatcher.

_process

Accumulating process mode is deprecated in VST 2.4! Use processReplacing instead!

setParameter

Set value of automatable parameter.

getParameter

Get value of automatable parameter.

numPrograms

Number of programs (Presets).

numParams

Number of parameters. All programs are assumed to have this many parameters.

numInputs

Number of audio inputs.

numOutputs

Number of audio outputs.

flags

Bitmask made of values from api::flags.

use vst2::api::flags;
let flags = flags::CAN_REPLACING | flags::CAN_DOUBLE_REPLACING;
// ...

reserved1

Reserved for host, must be 0.

reserved2

Reserved for host, must be 0.

initialDelay

For algorithms which need input in the first place (Group delay or latency in samples).

This value should be initially in a resume state.

_realQualities

Deprecated unused member.

_offQualities

Deprecated unused member.

_ioRatio

Deprecated unused member.

object

Void pointer usable by api to store object data.

user

User defined pointer.

uniqueId

Registered unique identifier (register it at Steinberg 3rd party support Web). This is used to identify a plug-in during save+load of preset and project.

version

Plug-in version (e.g. 1100 for v1.1.0.0).

processReplacing

Process audio samples in replacing mode.

processReplacingF64

Process double-precision audio samples in replacing mode.

future

Reserved for future use (please zero).

Methods

impl AEffect

unsafe fn get_plugin(&mut self) -> &mut Box<Plugin>

Return handle to Plugin object. Only works for plugins created using this library.

unsafe fn drop_plugin(&mut self)

Drop the Plugin object. Only works for plugins created using this library.