Struct vst2::api::Events

#[repr(C)]

pub struct Events {
    pub num_events: i32,
    pub _reserved: isize,
    pub events: [*mut Event; 2],
}

A struct which contains events.

Fields
num_events: i32

Number of events.

_reserved: isize

Reserved for future use. Should be 0.

events: [*mut Event; 2]

Variable-length array of pointers to api::Event objects.

The VST standard specifies a variable length array of initial size 2. If there are more than 2 elements a larger array must be stored in this structure.