Struct vst2::buffer::Outputs

pub struct Outputs<'a, T: 'a> { /* fields omitted */ }

Wrapper type to access the buffers for the output channels of an AudioBuffer in a safe way. Behaves like a slice.

Methods

impl<'a, T> Outputs<'a, T>

fn len(&self) -> usize

Number of channels

fn get(&self, i: usize) -> &'a [T]

Access channel at the given index, unchecked

fn get_mut(&self, i: usize) -> &'a mut [T]

Mutably access channel at the given index, unchecked

fn split_at_mut(&mut self, i: usize) -> (Outputs<'a, T>, Outputs<'a, T>)

Split borrowing at the given index, like for slices