Struct vst2::buffer::SendEventBuffer

pub struct SendEventBuffer { /* fields omitted */ }

This buffer is used for sending midi events through the VST interface. The purpose of this is to convert outgoing midi events from event::Event to api::Events. It only allocates memory in new() and reuses the memory between calls.

Methods

impl SendEventBuffer

fn new(capacity: usize) -> Self

Creates a buffer for sending up to the given number of midi events per frame

fn store<'a, T: IntoIterator<Item = U>, U: Borrow<Event<'a>>>(
&mut self,
events: T
)

Use this for sending events to a host or plugin.

Example

// Processor that clips samples above 0.4 or below -0.4:
fn process(&mut self, buffer: &mut AudioBuffer<f32>){
    let events = vec![
        // ...
    ];
    self.send_buffer.store(&events);
    self.host.process_events(self.send_buffer.events());
}

fn store_midi<T: IntoIterator<Item = U>, U: Borrow<MidiEvent>>(
&mut self,
events: T
)

Use this for sending midi events to a host or plugin. Like store() but for when you're not sending any SysExEvents, only MidiEvents.

fn events(&self) -> &Events

Use this for sending midi events to a host or plugin. See store()