Module vst2::channels

Meta data for dealing with input / output channels. Not all hosts use this so it is not necessary for plugin functionality.

Structs

ChannelInfo
Information about an input / output channel. This isn't necessary for a channel to function but informs the host how the channel is meant to be used.

Enums

ArrangementTarget
Target for Speaker arrangement type. Can be a cinema configuration or music configuration. Both are technically identical but this provides extra information to the host.

SpeakerArrangementType
Type representing how a channel is used. Only useful for some hosts.

StereoChannel
An enum for all channels in a stereo configuration.

StereoConfig
Possible stereo speaker configurations.

SurroundConfig
Possible surround speaker configurations.