Enum vst2::event::Event

pub enum Event<'a> {
    Midi {
        data: [u8; 3],
        delta_frames: i32,
        live: bool,
        note_length: Option<i32>,
        note_offset: Option<i32>,
        detune: i8,
        note_off_velocity: u8,
    },
    SysEx {
        payload: &'a [u8],
        delta_frames: i32,
    },
    Deprecated(Event),
}

A VST event.

Variants

Midi

A midi event.

These are sent to the plugin before Plugin::processing() or Plugin::processing_f64() is called.

Fields

data

The raw midi data associated with this event.

delta_frames

Number of samples into the current processing block that this event occurs on.

E.g. if the block size is 512 and this value is 123, the event will occur on sample samples[123].

live

This midi event was created live as opposed to being played back in the sequencer.

This can give the plugin priority over this event if it introduces a lot of latency.

note_length

The length of the midi note associated with this event, if available.

note_offset

Offset in samples into note from note start, if available.

detune

Detuning between -63 and +64 cents.

note_off_velocity

Note off velocity between 0 and 127.

SysEx

A system exclusive event.

This is just a block of data and it is up to the plugin to interpret this. Generally used by midi controllers.

Fields

payload

The SysEx payload.

delta_frames

Number of samples into the current processing block that this event occurs on.

E.g. if the block size is 512 and this value is 123, the event will occur on sample samples[123].

Deprecated

A deprecated event.

Passes the raw midi event structure along with this so that implementors can handle optionally handle this event.