Struct vst2::host::Host

pub struct Host {
    // some fields omitted
}

A reference to the host which allows the plugin to call back and access information.

Panics

All methods in this struct will panic if the plugin has not yet been initialized. In practice, this can only occur if the plugin queries the host for information when Default::default() is called.

struct ExamplePlugin;

impl Default for ExamplePlugin {
    fn default() -> ExamplePlugin {
        // Will panic, don't do this. If needed, you can query
        // the host during initialization via Vst::new()
        let host: Host = Default::default();
        let version = host.vst_version();

        // ...
    }
}

Methods

impl Host

fn automate(&mut self, index: i32, value: f32)

Notify the host that a parameter value was changed.

fn vst_version(&self) -> i32

Get the VST API version supported by the host e.g. 2400 = VST 2.4.

fn get_plugin_id(&self) -> i32

Get the plugin ID the host is requesting to load.

This is only useful for shell plugins where this value will change the plugin returned. TODO: implement shell plugins

fn idle(&self)

Tell the host that it can idle.

This is useful when the plugin is doing something such as mouse tracking in the UI.