Module vst2::plugin

Plugin specific structures.

Structs

HostCallback
A reference to the host which allows the plugin to call back and access information.

Info
A structure representing static plugin information.

Enums

CanDo
Features which are optionally supported by a plugin. These are queried by the host at run time.

Category
Plugin type. Generally either Effect or Synth.

Traits

Plugin
Must be implemented by all VST plugins.