Enum vst2::plugin::Category

#[repr(usize)]

pub enum Category {
    Unknown,
    Effect,
    Synth,
    Analysis,
    Mastering,
    Spacializer,
    RoomFx,
    SurroundFx,
    Restoration,
    OfflineProcess,
    Shell,
    Generator,
}

Plugin type. Generally either Effect or Synth.

Other types are not necessary to build a plugin and are only useful for the host to categorize the plugin.

Variants
Unknown

Unknown / not implemented

Effect

Any effect

Synth

VST instrument

Analysis

Scope, tuner, spectrum analyser, etc.

Mastering

Dynamics, etc.

Spacializer

Panners, etc.

RoomFx

Delays and Reverbs

SurroundFx

Dedicated surround processor.

Restoration

Denoiser, etc.

OfflineProcess

Offline processing.

Shell

Contains other plugins.

Generator

Tone generator, etc.