Struct vst2::plugin::HostCallback

pub struct HostCallback {
    // some fields omitted
}

A reference to the host which allows the plugin to call back and access information.

Panics

All methods in this struct will panic if the plugin has not yet been initialized. In practice, this can only occur if the plugin queries the host for information when Default::default() is called.

struct ExamplePlugin;

impl Default for ExamplePlugin {
    fn default() -> ExamplePlugin {
        // Will panic, don't do this. If needed, you can query
        // the host during initialization via Vst::new()
        let host: HostCallback = Default::default();
        let version = host.vst_version();

        // ...
    }
}

Methods

impl HostCallback

fn vst_version(&self) -> i32

Get the VST API version supported by the host e.g. 2400 = VST 2.4.