Struct vst2::Info

pub struct Info {
    pub name: String,
    pub vendor: String,
    pub presets: i32,
    pub parameters: i32,
    pub inputs: i32,
    pub outputs: i32,
    pub unique_id: i32,
    pub version: i32,
    pub category: PluginCategory,
    pub initial_delay: i32,
    pub preset_chunks: bool,
    pub f64_precision: bool,
}

A structure representing static plugin information.

Fields

name

Plugin Name.

vendor

Plugin Vendor.

presets

Number of different presets.

parameters

Number of parameters.

inputs

Number of inputs.

outputs

Number of outputs.

unique_id

Unique plugin ID. Can be registered with Steinberg to prevent conflicts with other plugins.

This ID is used to identify a plugin during save and load of a preset and project.

version

Plugin version (e.g. 0001 = v0.0.0.1, 1283 = v1.2.8.3).

category

Plugin category. Possible values are found in enums::PluginCategory.

initial_delay

preset_chunks

Indicates whether preset data is handled in formatless chunks. If false, host saves and restores plugins by reading/writing parameter data. If true, it is up to the plugin to manage saving preset data by implementing the {get, load}_{preset, bank}_chunks() methods. Default is false.

f64_precision

Indicates whether this plugin can process f64 based AudioBuffer buffers. Default is false.