Crate vst2

rust-vst2 is a rust implementation of the VST2.4 API

Plugins

All Plugins must implement the Vst trait and std::default::Default. The vst_main! macro must also be called in order to export the necessary functions for the VST to function.

Vst Trait

All methods in this trait have a default implementation except for the get_info method which must be implemented by the Vst object. Any of the default implementations may be overriden for custom functionality; the defaults do nothing on their own.

vst_main! macro

vst_main! will export the necessary functions to create a proper VST. This must be called with your VST struct name in order for the vst to work.

Example plugin

A barebones VST plugin:

#[macro_use]
extern crate vst2;
use std::default::Default;

use vst2::{Vst, Info};

#[derive(Default)]
struct BasicVst;

impl Vst for BasicVst {
    fn get_info(&self) -> Info {
        Info {
            name: "BasicVst".to_string(),

            ..Default::default()
        }
    }
}

vst_main!(BasicVst); //Important!

Hosts

Hosts are currently not supported. TODO

Reexports

pub use buffer::AudioBuffer;

Modules

api
Structures and types for interfacing with the VST 2.4 API.

buffer
Buffers to safely work with audio samples.

channels
Meta data for dealing with input / output channels. Not all hosts use this so it is not necessary for plugin functionality.

editor
All VST plugin editor related functionality.

enums
Enums for use in this library.

Macros

vst_main!
Exports the necessary symbols for the plugin to be used by a vst host.

Structs

Info
A structure representing static plugin information.

Traits

Vst
Must be implemented by all VST plugins.