Module vst2::api

Structures and types for interfacing with the VST 2.4 API.

Modules

consts
Constant values

Structs

AEffect
Used with the VST API to pass around plugin information.

ChannelProperties
Information about a channel. Only some hosts use this information.

Enums

SpeakerArrangementType
Tells the host how the channels are intended to be used in the plugin. Only useful for some hosts.

Type Definitions

DispatcherProc
Dispatcher function used to process opcodes. Called by host as a callback function.

GetParameterProc
Callback function used to get parameter values. Called by host.

HostCallback
Host callback function passed to VST. Can be used to query host information from plugin.

ProcessProc
Process function used to process 32 bit floating point samples.

ProcessProcF64
Process function used to process 64 bit floating point samples.

SetParameterProc
Callback function used to set parameter values. Called by host.