Struct vst2::buffer::AudioBuffer

pub struct AudioBuffer<'a, T: 'a + Float> { /* fields omitted */ }

AudioBuffer contains references to the audio buffers for all input and output channels

Methods

impl<'a, T: 'a + Float> AudioBuffer<'a, T>

fn new(
inputs: &'a [*const T],
outputs: &'a mut [*mut T],
samples: usize
) -> Self

Create an AudioBuffer from slices of raw pointers. Useful in a Rust VST host.

fn from_raw(
input_count: usize,
output_count: usize,
inputs_raw: *const *const T,
outputs_raw: *mut *mut T,
samples: usize
) -> Self

Create an AudioBuffer from raw pointers. Only really useful for interacting with the VST API.

fn input_count(&self) -> usize

The number of input channels that this buffer was created for

fn output_count(&self) -> usize

The number of output channels that this buffer was created for

fn samples(&self) -> usize

The number of samples in this buffer (same for all channels)

fn split<'b>(&'b mut self) -> (Inputs<'b, T>, Outputs<'b, T>) where
'a: 'b,

Split this buffer into separate inputs and outputs.

fn zip<'b>(&'b mut self) -> Zip<InputIterator<'b, T>, OutputIterator<'b, T>> where
'a: 'b,

Zip together buffers.