Enum vst2::channels::SpeakerArrangementType

pub enum SpeakerArrangementType {
    Custom,
    Empty,
    Mono,
    Stereo(StereoConfig, StereoChannel),
    Surround(SurroundConfig),
}

Type representing how a channel is used. Only useful for some hosts.

Variants
Custom

Custom arrangement not specified to host.

Empty

Empty arrangement.

Mono

Mono channel.

Stereo(StereoConfig, StereoChannel)

Stereo channel. Contains type of stereo arrangement and speaker represented.

Surround(SurroundConfig)

Surround channel. Contains surround arrangement and target (cinema or music).

Methods

impl SpeakerArrangementType

fn is_speaker_type(&self) -> bool

Determine whether this channel is part of a surround speaker arrangement.

fn is_left_stereo(&self) -> bool

Determine whether this channel is the left speaker in a stereo pair.