Enum vst2::channels::SurroundConfig

pub enum SurroundConfig {
    S3_0,
    S3_1,
    S4_0,
    S4_1,
    S5_0,
    S5_1,
    S6_0,
    S6_1,
    S7_0,
    S7_1,
    S8_0,
    S8_1,
    S10_2,
}

Possible surround speaker configurations.

Variants

S3_0

3.0 surround sound. Cinema: L R C Music: L R S

S3_1

3.1 surround sound. Cinema: L R C Lfe Music: L R Lfe S

S4_0

4.0 surround sound. Cinema: L R C S (LCRS) Music: L R Ls Rs (Quadro)

S4_1

4.1 surround sound. Cinema: L R C Lfe S (LCRS + Lfe) Music: L R Ls Rs (Quadro + Lfe)

S5_0

5.0 surround sound. Cinema and music: L R C Ls Rs

S5_1

5.1 surround sound. Cinema and music: L R C Lfe Ls Rs

S6_0

6.0 surround sound. Cinema: L R C Ls Rs Cs Music: L R Ls Rs Sl Sr

S6_1

6.1 surround sound. Cinema: L R C Lfe Ls Rs Cs Music: L R Ls Rs Sl Sr

S7_0

7.0 surround sound. Cinema: L R C Ls Rs Lc Rc Music: L R C Ls Rs Sl Sr

S7_1

7.1 surround sound. Cinema: L R C Lfe Ls Rs Lc Rc Music: L R C Lfe Ls Rs Sl Sr

S8_0

8.0 surround sound. Cinema: L R C Ls Rs Lc Rc Cs Music: L R C Ls Rs Cs Sl Sr

S8_1

8.1 surround sound. Cinema: L R C Lfe Ls Rs Lc Rc Cs Music: L R C Lfe Ls Rs Cs Sl Sr

S10_2

10.2 surround sound. Cinema + Music: L R C Lfe Ls Rs Tfl Tfc Tfr Trl Trr Lfe2