Struct vst2::channels::ChannelInfo

pub struct ChannelInfo {
    // some fields omitted
}

Information about an input / output channel. This isn't necessary for a channel to function but informs the host how the channel is meant to be used.

Methods

impl ChannelInfo

fn new(name: String, short_name: Option<String>, active: bool, arrangement_type: Option<SpeakerArrangementType>) -> ChannelInfo

Construct a new ChannelInfo. name is a user friendly name for this channel limited to MAX_LABEL characters. short_name is an optional field which provides a short name limited to MAX_SHORT_LABEL. active determines whether this channel is active. arrangement_type describes the arrangement type for this channel.

fn to_vst_api(self) -> ChannelProperties

Convert to the VST api equivalent of this structure.