Trait vst2::host::Host

pub trait Host {
    fn automate(&mut self, index: i32, value: f32) { ... }
    fn get_plugin_id(&self) -> i32 { ... }
    fn idle(&self) { ... }
    fn get_info(&self) -> (isize, String, String) { ... }
    fn process_events(&mut self, events: Vec<Event>) { ... }
}

Implemented by all VST hosts.

Provided Methods

fn automate(&mut self, index: i32, value: f32)

Automate a parameter; the value has been changed.

fn get_plugin_id(&self) -> i32

Get the plugin ID of the currently loading plugin.

This is only useful for shell plugins where this value will change the plugin returned. TODO: implement shell plugins

fn idle(&self)

An idle call.

This is useful when the plugin is doing something such as mouse tracking in the UI.

fn get_info(&self) -> (isize, String, String)

Get vendor and product information.

Returns a tuple in the form of (version, vendor_name, product_name).

fn process_events(&mut self, events: Vec<Event>)

Handle incoming events from the plugin.